Working with v1.2
Using the new app handlers now. I ran into a problem with those though.
I was using the master version of F2.js that I downloaded from GitHub
and I was getting a "type error: no method apply..." I switched to the
minified version that was being used in the JSFiddle example for
AppHandlers and that fixed it. Strange...